GWO
This algorithm is based on the leadership hierarchy of grey wolves and their hunting habits. Four distinct grey wolves alpha, beta, delta and omega (in hierarchical order) are employed to implement their hierarchy. There are three main steps in this algorithm, of which one is hierarchy management, followed by encircling prey and finally hunting.